pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nextferry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.6",
    "joblib>=1.3",
    "pandas>=2.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "scikit-learn>=1.4",
    "pytest>=8.0",
    "ruff>=0.9",
    "mypy>=1.14",
]

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end browser tests (require playwright and chromium)",
]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes (unused imports, undefined names)
    "I",    # isort (import ordering)
    "UP",   # pyupgrade (modern Python syntax)
    "B",    # flake8-bugbear (common bugs)
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long — handled by formatter for code, noisy for SQL/comments
]

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
ignore_missing_imports = true
# Start lenient — catch import errors and obvious issues in new code
# without requiring a full type annotation cleanup right now.
check_untyped_defs = false
disallow_untyped_defs = false
warn_return_any = false

# Existing modules with pre-existing type issues — tighten incrementally.
[[tool.mypy.overrides]]
module = [
    "backend.data_collector",
    "backend.database",
    "backend.display_processing",
    "backend.fill_predictor",
    "backend.ml_predictor",
    "backend.model_training.backtest_model",
    "backend.model_training.evaluation",
    "backend.model_training.report",
    "backend.next_sailings",
    "backend.serializers",
    "backend.utils",
    "backend.wsdot_client",
]
ignore_errors = true

[tool.setuptools.packages.find]
where = ["."]

[dependency-groups]
dev = [
    "pytest-playwright>=0.7.2",
]
